chore: move support.js and test.utils.js to test/e2e folder for synpress-specific eslint settings
