+++ package.json
@@ -84,8 +84,8 @@
     "vitest": "^0.32.0"
   },
   "resolutions": {
-    "**/@agoric/xsnap": "0.14.3-dev-9f085d3.0",
-    "**/@agoric/time": "0.3.3-dev-97cb715.0",
+    "**/@agoric/xsnap": "0.14.3-u14.0",
+    "**/@agoric/time": "0.3.3-u14.0",
     "**/@agoric/vats": "0.15.2-u15.0"
   }
 }
